test(e2e): centralize Discord gateway capture assertion
