+++ test/e2e/live/hermes-discord.test.ts
@@ -15,8 +15,11 @@ import { REPO_ROOT } from "../fixtures/paths.ts";
 import { buildProcessTokenProbe } from "../fixtures/process-token-probe.ts";
 import type { ShellProbeResult } from "../fixtures/shell-probe.ts";
 import { hermesDiscordHttpProxyWebSocketUrl } from "./hermes-discord-proxy.ts";
-import { type FakeDockerApi, startFakeDockerApi } from "./messaging-providers-helpers.ts";
-import { assertDiscordGatewayCapture } from "./openclaw-pairing-helpers.ts";
+import {
+  assertDiscordGatewayCapture,
+  type FakeDockerApi,
+  startFakeDockerApi,
+} from "./messaging-providers-helpers.ts";
 import {
   runSecondaryCleanup as bestEffortLifecycleCleanup,
   dockerInfo,
