+++ test/e2e/live/messaging-providers-helpers.ts
@@ -89,6 +89,23 @@ export type ChannelConfig = {
 export type AccountConfig = Record<string, unknown>;
 export { shellQuote };
 
+export function assertDiscordGatewayCapture(captureFile: string, expectedToken: string): void {
+  const rows = fs
+    .readFileSync(captureFile, "utf8")
+    .trim()
+    .split(/\n+/)
+    .filter(Boolean)
+    .map((line) => JSON.parse(line) as Record<string, unknown>);
+  const identify = rows.filter((row) => row.event === "identify").at(-1);
+  expect(identify, "fake Discord Gateway did not capture IDENTIFY").toBeTruthy();
+  expect(identify).not.toHaveProperty("token");
+  expect(JSON.stringify(rows), "fake Discord Gateway capture persisted raw token").not.toContain(
+    expectedToken,
+  );
+  expect(identify?.tokenMatchesExpected, "Discord token rewrite").toBe(true);
+  expect(identify?.tokenLooksPlaceholder, "Discord placeholder leaked").toBe(false);
+}
+
 export type FakeDockerApi = {
   kind: string;
   port: string;
